requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-Caching==1.10.1
Werkzeug==2.3.6
python-dotenv==1.0.0
